% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBson.R
\name{DBson}
\alias{DBson}
\title{DBson}
\usage{
DBson(slug)
}
\arguments{
\item{"slug"}{The slug provided on dbnomics}
}
\description{
This function allows you to easily download data formated in json from DBnomics and translate it to data frame
}
\examples{
# Download USA GDP from OECD's Economic Outlook
DBson("oecd-eo-usa-gdp-q")
}
\keyword{json}
