% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBson.R
\name{DBson}
\alias{DBson}
\title{DBson}
\usage{
DBson(slug, frequency = TRUE, source = FALSE)
}
\arguments{
\item{frequency}{A \emph{boolean}. If you want a frequency column in your data frame.}

\item{source}{A \emph{boolean}. If you want a data provider column in your data frame.}

\item{"slug"}{A \emph{string}. The slug provided on dbnomics.}
}
\value{
A \emph{long} data frame with at least 3 columns: \code{time values var}
}
\description{
This function allows you to easily download data formated in json from DBnomics and translate it to data frame
}
\examples{
# Download USA GDP from OECD's Economic Outlook
DBson("oecd-eo-usa-gdp-q")

}
\references{
\url{https://db.nomics.world/}
}
